agent/requirements.txt
openai==1.3.5
model-context-protocol>=0.1.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
dapr==1.12.0